% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Function to create a distance matrix}
\usage{
create_distance_matrix(score_file)
}
\arguments{
\item{score_file}{dataframe containing the score of pairwise alignment}
}
\value{
It returns a large matrix
}
\description{
This function creates a distance matrix from the pairwise alignment score file
}
\examples{

create_distance_matrix(score_file)
}
