% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Function to generate a tree based on the distance matrix provided}
\usage{
create_tree(distance_matrix, kmeans_output)
}
\arguments{
\item{distance_matrix}{distance matrix}

\item{kmeans_output}{k-means file}
}
\value{
Returns a tree with the sequences colored based on the cluster they belong to
}
\description{
This function creates a tree based on the distance matrix provided
and color the sequences based on the cluster they belong to
}
\examples{

create_tree(distance_matrix, kmeans_output)

}
