% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dim_reduced_file}
\alias{dim_reduced_file}
\title{Example of how should be the output of a dimension reduced file}
\format{
A dataframe containing rownames and 2 columns in the following order
\describe{
\item{axis_1}{position in axis 1}
\item{axis_2}{position in axis 2}
}
}
\usage{
dim_reduced_file
}
\description{
This database contains the position of each sequence in 2 dimensions
}
\keyword{datasets}
