% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Function to perform dimension reduction}
\usage{
dim_reduction(distance_matrix)
}
\arguments{
\item{distance_matrix}{distance matrix}
}
\value{
A dataframe containing the id, position in x axis, position in y axis
}
\description{
This function takes the distance matrix as input and reduce the dimension to 2 axis
}
\examples{

dim_reduction(distance_matrix)

}
