% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{distance_matrix}
\alias{distance_matrix}
\title{Example of how should be the output of the distance matrix file}
\format{
Large matrix
}
\usage{
distance_matrix
}
\description{
This database contains the distance matrix generated from BLAST file
}
\keyword{datasets}
