% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_score.R
\name{get_alignment_score}
\alias{get_alignment_score}
\title{Function to perform pairwise alignment of all sequences and retrieve the alignment score}
\usage{
get_alignment_score(fasta, type, alignment_method, cores = 2)
}
\arguments{
\item{fasta}{a fasta file containing all sequences from a multigene family}

\item{type}{if the sequences are from nucleotide (nuc) or protein (prot)}

\item{alignment_method}{method of sequence alignment (local) or (global)}

\item{cores}{number of cores to be used in the parallel computation}
}
\value{
It returns a dataframe containing the scores of each alignment
}
\description{
This function performs pairwise alignment of all sequences and compute the score of each alignment
}
\examples{

get_alignment_score(fasta, "prot", "global", 3)
}
