% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{get_variance}
\alias{get_variance}
\title{Function to get explained variance by each axis}
\usage{
get_variance(distance_matrix)
}
\arguments{
\item{distance_matrix}{distance matrix}
}
\value{
Returns a dataframe containing 1 observation and n variables, depending of the number of input sequences
}
\description{
This function gets the variance explained by each axis in the distance
matrix
}
\examples{

get_variance(distance_matrix)

}
