% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clustering.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{Function to estimate optimal number of clusters and clustering using kmeans}
\usage{
kmeans_clustering(dim_reduced_file, min_clust = 2, max_clust = 20)
}
\arguments{
\item{dim_reduced_file}{dimension reduced file}

\item{min_clust}{minimum number of clusters}

\item{max_clust}{maximum number of clusters}
}
\value{
A dataframe containing name of the seq, axis_1, axis_2 and cluster number)
}
\description{
This function estimates the optimal number of clusters in the input sample based on dimension reduced file. In case there are two or more cluster numbers supported by the same index number, the smaller number of clusters will be chosen to represent the family.
}
\examples{

kmeans_clustering(dim_reduced_file)

}
