% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_output.R
\docType{data}
\name{kmeans_output}
\alias{kmeans_output}
\title{Example of kmeans_output}
\format{
dataframe containing 4 columns: gene name, axis_1, axis_2, cluster
}
\usage{
kmeans_output
}
\description{
This is an example of the output of the kmeans_clustering
}
\examples{
# Load file
data(kmeans_output)

# Visualize the file
head(kmeans_output)
}
\keyword{datasets}
