% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations-meteo.R
\name{plot_stations_meteo}
\alias{plot_stations_meteo}
\title{plot_stations_meteo}
\usage{
plot_stations_meteo(file)
}
\arguments{
\item{file}{file location}
}
\value{
a graph
}
\description{
plot_stations_meteo
}
\examples{
\dontrun{
plot_stations_meteo(file = "https://data.toulouse-metropole.fr/explore/dataset/1-station-meteopole/download?format=csv&timezone=Europe/Berlin&use_labels_for_header=true")
}

}
