% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations-meteo.R
\name{read_station_meteo}
\alias{read_station_meteo}
\title{Read stations meteo}
\usage{
read_station_meteo(file)
}
\arguments{
\item{file}{}
}
\value{
a tibble
}
\description{
Read stations meteo
}
\examples{
read_station_meteo(
file = "https://data.toulouse-metropole.fr/explore/dataset/1-station-meteopole/download?format=csv&timezone=Europe/Berlin&use_labels_for_header=true"
)

}
