% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_file.R
\docType{data}
\name{score_file}
\alias{score_file}
\title{Example of score file}
\format{
Dataframe containing the score of each pairwise alignment in a wide format.
}
\usage{
score_file
}
\description{
This is an example of the ouput of the get_alignment_score function
}
\examples{
# Load file
data(score_file)

# Visualize the file
head(score_file)
}
\keyword{datasets}
