% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_cluster.R
\name{similarity_cluster}
\alias{similarity_cluster}
\title{Function to assess similarity within clusters}
\usage{
similarity_cluster(score_file, kmeans_output, output_type)
}
\arguments{
\item{score_file}{file containing the scores of each pairwise alignment}

\item{kmeans_output}{k-means file}

\item{output_type}{type of output (table or plot)}
}
\value{
It returns a dataframe or a ggplot figure
}
\description{
This function assess the mean/median similarity of each cluster
}
\examples{

similarity_cluster(score_file, kmeans_output, output_type = "plot")
similarity_cluster(score_file, kmeans_output, output_type = "table")

}
