% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/tcmuc_fasta.R
\docType{data}
\name{tcmuc_fasta}
\alias{tcmuc_fasta}
\title{Example of a complete fasta file that can be used as input}
\format{
complete fasta file containing headers and sequences
\describe{

}

fasta
}
\usage{
tcmuc_fasta

tcmuc_fasta
}
\description{
This database contains sequences in fasta format

This is an example of a fasta file from TcMUC family to be used as input
}
\examples{
# Load file
data(tcmuc_fasta)

# Visualize the file
head(tcmuc_fasta)
}
\keyword{datasets}
