% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/fasta.R
\docType{data}
\name{fasta}
\alias{fasta}
\title{Example of a small fasta file that can be used as input}
\format{
small fasta file containing headers and sequences
\describe{

}

fasta
}
\usage{
fasta

fasta
}
\description{
This database contains sequences in fasta format

This is an example of a fasta file to be used as input
}
\examples{
# Load file
data(fasta)

# Visualize the file
head(fasta)
}
\keyword{datasets}
